+++ changelog.md
@@ -1,3 +1,12 @@
+0.2.3
+-----
+
+- Fix typo on polling timeline in answer module.
+- Raise NoAnswerError when there is no matching grams.
+- Get mentions when threshold is set.
+- Compatibility with tweepy>=3.0.
+
+
 0.2.2
 -----